requirements.txt
azure-functions==1.21.3
azure-identity==1.19.0
azure-storage-blob==12.24.0
pytest==8.3.4
requests==2.32.3

